addition of check_snpeff_errors and check_snpeff_warnings functions
check_snpeff_warnings and check_snpeff_errors reads in the snpeff annotations and will return (if present) the number of warning messages and their types, and the number of error messages and their types. Warning messages do not interefere with prewas, so the prewas run is allowed to continue. Errors indicate a problem with the annotation, and need to be cleaned up prior to prewas, so cause prewas to stop.